[refact]useTheme 리팩토링
- useThemeEffect 훅스 삭제.
- useTheme 훅으로 병합
